requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0
pandas>=1.3.0
openpyxl>=3.0.0
pydantic>=2.0.0
PyYAML>=6.0
requests>=2.25.0